Fix: Do not define 2 versions of freelist_iter_tpl
DEBUG_FREELIST was defined in freelist_tpl.h but not in freelist_iter_tpl.h,
causing two versions of freelist_iter_tpl to be emitted, causing a compile error with -flto
and possibly some crashes in versions without -flto, too.

Define DEBUG_FREELIST in simconst.h to have a single place to enable
the compile-time switch.

git-svn-id: svn://servers.simutrans.org/simutrans/trunk@11511 8aca7d54-2c30-db11-9de9-000461428c89